update drivers license

+++ backend/src/main/java/com/logumrh/repository/common/DriversLicenseRepository.java
@@ -3,6 +3,6 @@ package com.logumrh.repository.common;
 import com.logumrh.model.common.DriversLicense;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DriversLicenseRepository extends JpaRepository<DriversLicense, Integer> {
+public interface DriversLicenseRepository extends JpaRepository<DriversLicense, Long> {
 
 }
